function/requirements.txt
functions-framework==3.5.0
google-cloud-compute==1.19.0
google-cloud-firestore==2.16.0
